pyproject.toml
[tool.poetry]
name = "scenarios-conductor"
version = "0.1.0"
description = ""
authors = ["Ruslan Babayev <rus.babaef@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
structlog = "^25.4.0"
pydantic = "^2.11.5"
python-dotenv = "^1.1.0"
aiohttp = "^3.12.9"
prometheus-client = "^0.22.1"
otteroad = "^0.1.7"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.7"
isort = "^6.0.1"
black = "^25.1.0"
pytest = {version = "^8.4.0", extras = ["asyncio"]}
pytest-cov = "^6.1.1"
pytest-asyncio = "^1.0.0"
pytest-coverage = "^0.0"

[tool.poetry.scripts]
launch_app = "scenarios_conductor.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
split_on_trailing_comma = true
profile = "black"
